# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API",
        "description": "Swagger documentation for the Job Portal service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get all job listings",
                "operationId": "7ddb43f87de5ddc0eb46ffea0b1ce86c",
                "responses": {
                    "200": {
                        "description": "List of jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            },
                                            "location": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/public/jobs": {
            "get": {
                "tags": [
                    "Public Jobs"
                ],
                "summary": "Get all job listings (public)",
                "description": "Public endpoint - List job tanpa autentikasi (read-only)",
                "operationId": "57bd29ec0f32fb3b8e4d9ff73c29e9a8",
                "responses": {
                    "200": {
                        "description": "List of jobs"
                    }
                }
            }
        },
        "/api/public/jobs/{id}": {
            "get": {
                "tags": [
                    "Public Jobs"
                ],
                "summary": "Get job detail (public)",
                "description": "Public endpoint - Detail 1 job tanpa autentikasi",
                "operationId": "90150990e17240fa4c448681e31da812",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job detail"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Jobs",
            "description": "Jobs"
        },
        {
            "name": "Public Jobs",
            "description": "Public Jobs"
        }
    ]
}